+++ src/main/webapp/components/lists/index.js
